modified:   Resources/mediacastproducer/lib/mcp/actions.rb 	modified:   Resources/mediacastproducer/lib/mcp/actions/toolscript.rb 	new file:   Resources/mediacastproducer/lib/mcp/common/asl_class_logging.rb 	modified:   Resources/mediacastproducer/lib/mcp/plist/propertylist.rb 	modified:   Resources/mediacastproducer/lib/mcp/plist/script_template.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/base.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/ffmpeg.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/ffmpeg2theora.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/mediastreamsegmenter.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/mp4faststart.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/streamsegmenter.rb 	modified:   Resources/mediacastproducer/lib/mcp/tools/vlc.rb 	modified:   Resources/mediacastproducer/lib/mcp/transcoder.rb 	new file:   Resources/mediacastproducer/lib/mcp/transcoder/presets/vlc_mp4_cbr.plist 	new file:   Resources/mediacastproducer/lib/mcp/transcoder/presets/vlc_mp4_cbr_3:2_480x320_576kbit_44,1khz_stereo_192kbit.plist 	new file:   Resources/mediacastproducer/lib/mcp/transcoder/presets/vlc_mpeg2_ts_mediastreamsegmenter.plist 	copied:     Resources/mediacastproducer/lib/mcp/transcoder/scripts/vlc_ogv_cbr.plist -> Resources/mediacastproducer/lib/mcp/transcoder/scripts/vlc_mp4_cbr.plist 	new file:   Resources/mediacastproducer/lib/mcp/transcoder/scripts/vlc_mpeg2_ts_mediastreamsegmenter.plist 	modified:   Resources/mediacastproducer/lib/mcp/transcoder/scripts/vlc_ogv_cbr.plist 	modified:   Resources/mediacastproducer/lib/mcp/transcoder/scripts/vlc_ogv_qbr.plist 	modified:   Resources/mediacastproducer/lib/mcp/transcoder/scripts/vlc_ogv_vbr.plist

+++ Resources/mediacastproducer/lib/mcp/actions/toolscript.rb
@@ -126,30 +126,42 @@ module PodcastProducer
           tools << tool
         end
         begin
-          arguments = MediacastProducer::Common::TemplateArray.substitute(@script.arguments + arguments, data)
+          pipeline = []
+          preset.script.commands.each do |cmd|
+            pipeline << MediacastProducer::Common::TemplateArray.substitute(preset.script.arguments[cmd], data)
+          end
+          pipeline << arguments unless arguments.empty?
         rescue McastTemplateException => e
           log_crit_and_exit(e.message,e.return_code.to_i)
         end
-        log_notice(arguments.join(' '))
-        unless (pid = fork_exec_and_return_pid(*arguments))
+        cmds = []
+        pipeline.each { |args|  cmds << args.join(' ') }
+        log_notice(cmds.join(' | '))
+        exit
+        unless (pids = fork_chain_and_return_pids(*pipeline))
           log_crit_and_exit("failed to transcode '#{@input}' to '#{@output}' with script '#{@script_name}'", -1)
         end
         begin
           puts "<xgrid>\n{control = statusUpdate; percentDone = 0.0; }\n</xgrid>"
           tools.each do |tool|
             next unless tool.respond_to?(:update_status)
-            tool.update_status(pid) { |percent|
+            tool.update_status { |percent|
               puts "<xgrid>\n{control = statusUpdate; percentDone = #{percent}; }\n</xgrid>"
             }
           end
-          pid, status = Process.waitpid2(pid)
-          puts "<xgrid>\n{control = statusUpdate; percentDone = 100.0; }\n</xgrid>"
+          force_quit = false
         rescue SystemExit, Interrupt
-          Process.kill('HUP', pid)
-          pid, status = Process.waitpid2(pid)
+          force_quit = true
         end
-        log_notice("pid: #{pid} exit status: #{status.exitstatus}")
-        log_crit_and_exit("failed to transcode '#{@input}' to '#{@output}' with script '#{@script_name}'",status.exitstatus) unless status.exitstatus == 0
+        result = true
+        pids.each do |pid|
+          Process.kill('HUP', pid) if force_quit
+          pid, status = Process.waitpid2(pid)
+          log_notice("pid: #{pid} exit status: #{status.exitstatus}")
+          result = false unless status.exited? && status.exitstatus == 0
+        end
+        puts "<xgrid>\n{control = statusUpdate; percentDone = 100.0; }\n</xgrid>" if result && !force_quit
+        log_crit_and_exit("failed to transcode '#{@input}' to '#{@output}' with script '#{@script_name}'", -1) unless result
       end
     end
